+++ pkg/proxy/utils_test.go
@@ -6,7 +6,6 @@ import (
 	"sync"
 
 	"github.com/nothinux/octo-proxy/pkg/config"
-	"github.com/nothinux/octo-proxy/pkg/tlsconn"
 )
 
 func SendData(hc config.HostConfig, message []byte, readResponse bool) error {
@@ -37,7 +36,7 @@ func SendData(hc config.HostConfig, message []byte, readResponse bool) error {
 }
 
 func RunTestTLSServer(wg *sync.WaitGroup, c config.TLSConfig, result chan []byte) string {
-	tlsConfig, err := tlsconn.GetTLSConfig(c)
+	tlsConfig, err := getTLSConfig(c)
 	if err != nil {
 		log.Fatal(err)
 	}
